Reformat comment for readability, 80-col rule, and gcc cleanliness
Summary:
The continuation at the end of the line with the while
was causing gcc to complain about it because it's a
single line comment.

Reformat it into a multiline comment and wrap it to 80 chars.

Test Plan: It's just a comment...

Reviewed By: paulbiss@fb.com

Subscribers: trunkagent, folly-diffs@, yfeldblum

FB internal diff: D1850876

Signature: t1:1850876:1424124973:b2afbc2a8177ac6d11a21bf5a52093bcfd9cbb3e