then-with-Executor
Summary:
Pass an Executor to `then`, which applies only for that callback. This is on
the one hand just a convenience method, but it's a major convenience when it's
needed, because grabbing, storing, and restoring the old Executor in the
middle of a chain is very inconvenient indeed.

Test Plan:
new unit

Reviewed By: jsedgwick@fb.com

Subscribers: robbert, exa, folly-diffs@, jsedgwick, nkgupta, yfeldblum, chalfant, davejwatson

FB internal diff: D1985475

Signature: t1:1985475:1429148056:0450120263ba8110e8825420cbefe3b1887f7306