then-with-Executor
Summary:
Pass an Executor to `then`, which applies only for that callback. This is on
the one hand just a convenience method, but it's a major convenience when it's
needed, because grabbing, storing, and restoring the old Executor in the
middle of a chain is very inconvenient indeed.

Test Plan:
new unit
docblock comments

Reviewed By: jsedgwick@fb.com

Subscribers: folly-diffs@, davejwatson, chalfant, yfeldblum, nkgupta, jsedgwick, exa, robbert

FB internal diff: D2011542

Tasks: 6771589, 6838553

Signature: t1:2011542:1429660204:f5959b1e0b3b36dfb8c3c7091302d19101dde93b