Get *=default*ed default constructors
Summary: Defaulted (*=default*ed) default constructors are better

because they can be constexpr and/or noexcept when C++ Standard

thinks it is right to do so. And they remain user-declared rather

than user-provided. Regarding *=default*ed default constructor,

benifit is that the work might be done during compilation and we

might not need to worry about exception paths. And for destructors,

apart from that that =defaulted definition is in sync with defaulted

default constructor we might discover that in some cases "() {}" might

be ill-formed when "= default;" compiltion indicates so. If =defaulted

definition for destructor doesn't do any harm then why not go for it.

Test Plan:

All folly/tests, make check for 37 tests, passed.