Remove double-conversion as a source dep in OS X script
Summary: double-conversion was recently added to homebrew, making it the last
folly dependency to do so. this diff modifies the existing OS X
bootstrap tool to account for that. I also added an rm for the gtest
wget, so there wasn't a leftover zip file in your repo. alternatively,
we could remove that and add it to the gitignore.

Reviewed By: @fugalh

Differential Revision: D2196000

Pulled By: @sgolemon