make folly/detail/FunctionalExcept.* work if FOLLY_HAVE_BITS_FUNCTEXCEPT_H is set or not
Summary: When doing android development, we need to compile with
FOLLY_HAVE_BITS_FUNCTEXCEPT_H set for some flavors, and
FOLLY_HAVE_BITS_FUNCTEXCEPT_H unset for others (primarily for tests
run in the local (xcode) development platform).  This diff adds some
more conditionals so it all works.

In this repro, these files are never used, so this is more or less a
noop, but this is the upstream so I'm making these changes here.

Reviewed By: @Gownta

Differential Revision: D2202107