Adjust Format-inl.h to work with MSVC
This adjusts a format string to be sane, I don't think `'` is a valid format specifier to begin with, and MSVC2015 added a warning for invalid specifiers, which this triggers.
This also adjusts a use of `std::enable_if` to eliminate the `enable_if` entirely, because MSVC doesn't like it when you try to do a sizeof check on a type like this. (I haven't checked, but I wouldn't be surprised if MSVC were just returning `-1` for the size here, either way, the new way it's written should still have the desired effect)