Port fbstring_core to big-endian architectures.
fbstring_core contains an implementation of the Small String Optimization.
The SSO is used to stack-allocate the string contents if the string is "small",
and heap-allocate the string contents if the string is "big". Specifically,
fbstring_core has "small", "medium", and "large" string categories for various
levels of optimizations.

The existing implementation of the SSO is little-endian dependent, where the
2 bits determining storage category ("small", "medium", or "large") are stored
in the 2 MSb of a size_t. On little-endian architectures, these 2 bits end up
in the far right byte of the struct, and as such the left-most bytes of the
size_t can be used reused during the SSO to hold actual string content. On
big-endian architectures, though, the 2 MSb of a size_t end up partway through
where the SSO wants to be storing actual string content. To make this work on
big-endian, we can simply use the 2 LSb of the size_t to encode the storage
category, as they'll be in the far right byte of the struct as we want. We then
need to simply shift our size_t value in order to interpret it correctly, rather
than masking off the category discriminators as on little-endian.