Add a compatibility shim for working with libevent on MSVC
MSVC builds of libevent expect `evutil_socket_t` to be `HANDLE` values, but Folly, HHVM, and Thrift all use them as file descriptors.
This adds a shim for the only method that needs it, `event_set`, and includes it where it's needed in Folly.
The shim is only enabled under MSVC, but it's safe to include the header from any platform.