Remove `unsigned char v = static_cast<unsigned char>(v);`
Summary: MSVC spotted this very suspicious line when I was running it with /analyze, and complained about using an unitialized variable. This variable isn't used anywhere in the function, and is initializing itself with itself. I have no idea how this compiled in the first place, but apparently it does.
This removes the line in question.
Closes #295

Reviewed By: @yfeldblum, @paulbiss

Differential Revision: D2340688

Pulled By: @sgolemon